Moved FlexFacetPropertyTester to c.g.c.t.e.util, renamed it to FacetExistsPropertyTester and made the functionality more general so that it supports any facet
