+++ plugins/com.google.cloud.tools.eclipse.appengine.ui/plugin.xml
@@ -50,8 +50,9 @@
               </parameter>
               <visibleWhen>
                   <test
-                  property="com.google.cloud.tools.eclipse.appengine.flex.flexFacetExists"
-                  forcePluginActivation="true"/>
+                        forcePluginActivation="true"
+                        property="com.google.cloud.tools.eclipse.util.facetExists"
+                        value="com.google.cloud.tools.eclipse.appengine.facets.flex"/>
               </visibleWhen>
            </command>
            <command
