+++ plugins/com.google.cloud.tools.eclipse.util/META-INF/MANIFEST.MF
@@ -28,5 +28,6 @@ Import-Package: com.google.common.annotations;version="[20.0.0,21.0.0)",
 Bundle-ClassPath: .,
  templates/appengine/,
  lib/google-http-client-1.22.0.jar
-Require-Bundle: org.eclipse.m2e.maven.runtime,
+Require-Bundle: org.eclipse.core.expressions,
+ org.eclipse.m2e.maven.runtime,
  org.eclipse.m2e.core;bundle-version="1.6.2"
